Fix #27
Uses Objects.nunNull to check if a value is not null.
Using this method, if the value given is a primitive type, it will work as well.

If we try to check if a `primitive != null` directly, it doesn't compile.

If the field is a primitive value, we could just skip the null check.
But this field being checked is the id of an object.
The RecordGenerator just has easy access to that object,
not its id attribute to check if it is a primitive value and then skip the null check.

This way, it's easier to use a null-check that safely works for primitive and non-primitives.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>